+++ examples/app_multi_node/train_lite.py
@@ -2,7 +2,7 @@ import torch
 
 import lightning as L
 from lightning.app.components import LiteMultiNode
-from lightning.lite import LightningLite
+from lightning.lite import Fabric
 
 
 class LitePyTorchDistributed(L.LightningWork):
@@ -14,8 +14,8 @@ class LitePyTorchDistributed(L.LightningWork):
             torch.nn.Linear(1, 1),
         )
 
-        # 2. Create LightningLite.
-        lite = LightningLite(strategy="ddp", precision=16)
+        # 2. Create Fabric.
+        lite = Fabric(strategy="ddp", precision=16)
         model, optimizer = lite.setup(model, torch.optim.SGD(model.parameters(), lr=0.01))
         criterion = torch.nn.MSELoss()