+++ src/lightning_lite/wrappers.py
@@ -45,7 +45,7 @@ class _LiteOptimizer:
         # `__del__` is skipped in case the optimizer has implemented custom destructor logic which we would
         # not want to call on destruction of the `_LiteOptimizer
         self.__dict__ = {k: v for k, v in optimizer.__dict__.items() if k not in ("state_dict", "step", "__del__")}
-        self.__class__ = type("Lite" + optimizer.__class__.__name__, (self.__class__, optimizer.__class__), {})
+        self.__class__ = type("Fabric" + optimizer.__class__.__name__, (self.__class__, optimizer.__class__), {})
         self._optimizer = optimizer
         self._strategy = strategy
 
@@ -82,7 +82,7 @@ class _LiteModule(_DeviceDtypeModuleMixin):
             forward_module: The module to wrap the ``forward`` method on.
             precision: Reference to the precision plugin for handling precision context
             original_module: The original, unmodified module as passed into the
-                :meth:`pytorch_lightning.lite.lite.LightningLite.setup` method. This is needed when attribute lookup
+                :meth:`pytorch_lightning.lite.lite.Fabric.setup` method. This is needed when attribute lookup
                 on this wrapper should pass through to the original module.
         """
         super().__init__()