+++ tests/tests_lite/test_connector.py
@@ -849,30 +849,26 @@ def test_arguments_from_environment_collision():
     """Test that the connector raises an error when the CLI settings conflict with settings in the code."""
     with mock.patch.dict(os.environ, {"LT_ACCELERATOR": "cpu"}):
         with pytest.raises(
-            ValueError, match=escape("Your code has `LightningLite(accelerator='cuda', ...)` but it conflicts")
+            ValueError, match=escape("Your code has `Fabric(accelerator='cuda', ...)` but it conflicts")
         ):
             _Connector(accelerator="cuda")
 
     with mock.patch.dict(os.environ, {"LT_STRATEGY": "ddp"}):
         with pytest.raises(
-            ValueError, match=escape("Your code has `LightningLite(strategy='ddp_spawn', ...)` but it conflicts")
+            ValueError, match=escape("Your code has `Fabric(strategy='ddp_spawn', ...)` but it conflicts")
         ):
             _Connector(strategy="ddp_spawn")
 
     with mock.patch.dict(os.environ, {"LT_DEVICES": "2"}):
-        with pytest.raises(ValueError, match=escape("Your code has `LightningLite(devices=3, ...)` but it conflicts")):
+        with pytest.raises(ValueError, match=escape("Your code has `Fabric(devices=3, ...)` but it conflicts")):
             _Connector(devices=3)
 
     with mock.patch.dict(os.environ, {"LT_NUM_NODES": "3"}):
-        with pytest.raises(
-            ValueError, match=escape("Your code has `LightningLite(num_nodes=2, ...)` but it conflicts")
-        ):
+        with pytest.raises(ValueError, match=escape("Your code has `Fabric(num_nodes=2, ...)` but it conflicts")):
             _Connector(num_nodes=2)
 
     with mock.patch.dict(os.environ, {"LT_PRECISION": "16"}):
-        with pytest.raises(
-            ValueError, match=escape("Your code has `LightningLite(precision=64, ...)` but it conflicts")
-        ):
+        with pytest.raises(ValueError, match=escape("Your code has `Fabric(precision=64, ...)` but it conflicts")):
             _Connector(precision=64)