+++ tests/tests_lite/test_wrappers.py
@@ -18,12 +18,12 @@ import torch
 from tests_lite.helpers.runif import RunIf
 from torch.utils.data.dataloader import DataLoader
 
-from lightning_lite.lite import LightningLite
+from lightning_lite.lite import Fabric
 from lightning_lite.utilities.device_dtype_mixin import _DeviceDtypeModuleMixin
 from lightning_lite.wrappers import _LiteDataLoader, _LiteModule, _LiteOptimizer
 
 
-class EmptyLite(LightningLite):
+class EmptyLite(Fabric):
     def run(self):
         pass